[Refactor] map_view.py -> frontend/streamlit_prototype/components로 이동
